extract kotlin http and sse
