+++ examples/cat_facts/Android/build.gradle
@@ -1,7 +1,6 @@
 buildscript {
     ext {
         compose_version = '1.3.1'
-        ktor_version = '2.2.2'
     }
 }
